Split worker into files (#9)
* Split worker into utils and core file

* Remove unnecessary package

* Fixed ajv import
