+++ package.json
@@ -51,7 +51,6 @@
   "dependencies": {
     "comlink": "^4.3.0",
     "lodash-es": "^4.17.15",
-    "object-sizeof": "^1.6.1",
     "uuid": "^8.3.1"
   },
   "files": [
