+++ src/hux-ql/application/compileAjvSchema.js
@@ -1,4 +1,4 @@
-import Ajv from "ajv";
+const Ajv = require("ajv");
 
 const compileAjvSchema = ({ schema }) => {
   const ajv = new Ajv();
